Fix auto publish script once and forall
