+++ scripts/publish-process/constants.js
@@ -25,7 +25,7 @@ const YES_REGEX = /Y(es)?/i;
 const YES_NO_REGEX = /Y(es)?|N(o)?/i;
 const CHANGELOG_PATH = 'CHANGELOG.md';
 const UNRELEASED_PATH = 'UNRELEASED.yaml';
-const UNRELEASED_PATH_CLEAN = 'scripts/release-process/UNRELEASED_CLEAN.yaml';
+const UNRELEASED_PATH_CLEAN = 'scripts/publish-process/UNRELEASED_CLEAN.yaml';
 const RELEASE_MODE_PATCH = 'patch';
 const RELEASE_MODE_MINOR = 'minor';
 const RELEASE_MODE_MAJOR = 'major';
